+++ core/src/index/thirdparty/faiss/utils/distances.h
@@ -188,7 +188,6 @@ void knn_jaccard (
         size_t d, size_t nx, size_t ny,
         float_maxheap_array_t * res,
         ConcurrentBitsetPtr bitset = nullptr);
-
 /** same as knn_L2sqr, but base_shift[bno] is subtracted to all
  * computed distances.
  *