+++ core/src/wrapper/BinVecImpl.cpp
@@ -16,9 +16,7 @@
 // under the License.
 
 #include "wrapper/BinVecImpl.h"
-
-#include <src/index/knowhere/knowhere/index/vector_index/IndexBinaryIVF.h>
-
+#include "index/knowhere/knowhere/index/vector_index/IndexBinaryIVF.h"
 #include "WrapperException.h"
 #include "knowhere/adapter/VectorAdapter.h"
 #include "knowhere/common/Exception.h"
@@ -270,6 +268,5 @@ BinBFIndex::AddWithoutIds(const int64_t& nb, const uint8_t* xb, const Config& cf
     std::static_pointer_cast<knowhere::BinaryIDMAP>(index_)->AddWithoutId(ret_ds, cfg);
     return Status::OK();
 }
-
 }  // namespace engine
 }  // namespace milvus