+++ core/unittest/db/test_mem.cpp
@@ -84,6 +84,7 @@ TEST_F(MemManagerTest, VECTOR_SOURCE_TEST) {
     milvus::engine::VectorSource source(vectors);
 
     size_t num_vectors_added;
+
     //    milvus::engine::ExecutionEnginePtr execution_engine_ = milvus::engine::EngineFactory::Build(
     //        table_file_schema.dimension_, table_file_schema.location_,
     //        (milvus::engine::EngineType)table_file_schema.engine_type_,
@@ -93,6 +94,7 @@ TEST_F(MemManagerTest, VECTOR_SOURCE_TEST) {
     auto segment_writer_ptr = std::make_shared<milvus::segment::SegmentWriter>(directory);
 
     status = source.Add(segment_writer_ptr, table_file_schema, 50, num_vectors_added);
+
     ASSERT_TRUE(status.ok());
     ASSERT_EQ(num_vectors_added, 50);
     ASSERT_EQ(source.GetVectorIds().size(), 50);
@@ -153,7 +155,6 @@ TEST_F(MemManagerTest, MEM_TABLE_TEST) {
     BuildVectors(n_100, vectors_100);
 
     milvus::engine::VectorSourcePtr source_100 = std::make_shared<milvus::engine::VectorSource>(vectors_100);
-
     milvus::engine::MemTable mem_table(GetTableName(), impl_, options);
 
     milvus::engine::IDNumbers vector_ids;
@@ -171,6 +172,7 @@ TEST_F(MemManagerTest, MEM_TABLE_TEST) {
     BuildVectors(n_max, vectors_128M);
 
     vector_ids.clear();
+
     milvus::engine::VectorSourcePtr source_128M = std::make_shared<milvus::engine::VectorSource>(vectors_128M);
     status = mem_table.Add(source_128M);
     ASSERT_TRUE(status.ok());
@@ -249,6 +251,7 @@ TEST_F(MemManagerTest2, SERIAL_INSERT_SEARCH_TEST) {
         std::vector<std::string> tags;
         milvus::engine::ResultIds result_ids;
         milvus::engine::ResultDistances result_distances;
+
         stat = db_->Query(dummy_context_, GetTableName(), tags, topk, nprobe, search, result_ids, result_distances);
         ASSERT_EQ(result_ids[0], pair.first);
         ASSERT_LT(result_distances[0], 1e-4);